// src/utils/types/tiers.interface.ts
export interface ITiersPrevisions {
  auteurSaisie: string;
  auteurStatut: string;
  avecTva: string;
  cle: string;
  cleAccordPartenaire: string;
  cleBudget: string;
  cleCommande: string;
  cleCourrier: string;
  cleEcriture: string;
  cleEtale: string;
  cleFacture: string;
  clePrelevement: string;
  cleRubriqueTreso: string;
  codeJournal: string;
  codeOperation: string;
  codeTiersPayeur: string;
  commentaire: string;
  compteRegroupement: string;
  credit: string;
  dateEcheance: string;
  dateOrdo: string;
  datePaiement: string;
  datePiece: string;
  dateSaisie: string;
  dateSolde: string;
  debit: string;
  dhStatut: string;
  inter: string;
  libelleCompteGeneral: string;
  libelleCompteTiers: string;
  libelleEcriture: string;
  libelleEcritureAnnee: string;
  libelleEcritureBeneficiaire: string;
  libelleEcritureMois: string;
  libelleEcriturePrefixe: string;
  libelleEcritureTrimestre: string;
  manu: string;
  marque: string;
  matMark: string;
  modeReglement: string;
  montantTva1: string;
  montantTva2: string;
  montantTva3: string;
  montantTva4: string;
  montantTva5: string;
  noCompteBanque: string;
  noCompteGeneral: string;
  noCompteTiers: string;
  noSemaineEcheance: string;
  noSemaineOrdo: string;
  refSourceTiers: string;
  referencePaiement: string;
  rubriqueTreso: string;
  societe: string;
  statut: string;
  tauxTva1: string;
  tauxTva2: string;
  tauxTva3: string;
  tauxTva4: string;
  tauxTva5: string;
}